+++ test/gtest/ucp/test_ucp_rma.cc
@@ -40,7 +40,7 @@ public:
                   reply_buffer);
     }
 
-    virtual unsigned features() const {
+    virtual uint64_t features() const {
         return UCP_FEATURE_RMA;
     }
 };
